requirements.txt
Flask
gunicorn
google-auth
google-auth-oauthlib
google-api-python-client
httpx[http2]
beautifulsoup4
google-cloud-storage
google-generativeai
telegraph
grpcio>=1.62.0
feedparser
cloudscraper
requests
